+++ src/repository/cloud_image.py
@@ -201,7 +201,7 @@ async def create_qr(body: ImageTransformModel, db: Session, user: User):
             status_code=status.HTTP_404_NOT_FOUND, detail=messages.IMAGE_NOT_FOUND
         )
     if image.qr_url:
-        return ImageQRResponse(image_id=image.id, qr_code_url="")
+        return ImageQRResponse(image_id=image.id, qr_code_url=image.qr_url)
     
     qr = qrcode.QRCode()
     qr.add_data(image.url)
